third_party/blink/web_tests/fast/dom/event-attribute-availability-expected.txt

This tests what event handler attributes are available on what objects.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Test Element
PASS 'onabort' in element is true
PASS 'onblur' in element is true
PASS 'onchange' in element is true
PASS 'onclick' in element is true
PASS 'oncontextmenu' in element is true
PASS 'ondblclick' in element is true
PASS 'ondrag' in element is true
PASS 'ondragend' in element is true
PASS 'ondragenter' in element is true
PASS 'ondragleave' in element is true
PASS 'ondragover' in element is true
PASS 'ondragstart' in element is true
PASS 'ondrop' in element is true
PASS 'onerror' in element is true
PASS 'onfocus' in element is true
PASS 'oninput' in element is true
PASS 'onkeydown' in element is true
PASS 'onkeypress' in element is true
PASS 'onkeyup' in element is true
PASS 'onload' in element is true
PASS 'onmousedown' in element is true
PASS 'onmousemove' in element is true
PASS 'onmouseout' in element is true
PASS 'onmouseover' in element is true
PASS 'onmouseup' in element is true
PASS 'onmousewheel' in element is true
PASS 'onscroll' in element is true
PASS 'onselect' in element is true
PASS 'onsubmit' in element is true
PASS 'onbeforecut' in element is true
PASS 'oncut' in element is true
PASS 'onbeforecopy' in element is true
PASS 'oncopy' in element is true
PASS 'onbeforepaste' in element is true
PASS 'onpaste' in element is true
PASS 'onreset' in element is true
PASS 'onsearch' in element is true
PASS 'onselectstart' in element is true

Test Document
PASS 'onabort' in document is true
PASS 'onblur' in document is true
PASS 'onchange' in document is true
PASS 'onclick' in document is true
PASS 'oncontextmenu' in document is true
PASS 'ondblclick' in document is true
PASS 'ondrag' in document is true
PASS 'ondragend' in document is true
PASS 'ondragenter' in document is true
PASS 'ondragleave' in document is true
PASS 'ondragover' in document is true
PASS 'ondragstart' in document is true
PASS 'ondrop' in document is true
PASS 'onerror' in document is true
PASS 'onfocus' in document is true
PASS 'oninput' in document is true
PASS 'onkeydown' in document is true
PASS 'onkeypress' in document is true
PASS 'onkeyup' in document is true
PASS 'onload' in document is true
PASS 'onmousedown' in document is true
PASS 'onmousemove' in document is true
PASS 'onmouseout' in document is true
PASS 'onmouseover' in document is true
PASS 'onmouseup' in document is true
PASS 'onmousewheel' in document is true
PASS 'onscroll' in document is true
PASS 'onselect' in document is true
PASS 'onsubmit' in document is true
PASS 'onbeforecut' in document is true
PASS 'oncut' in document is true
PASS 'onbeforecopy' in document is true
PASS 'oncopy' in document is true
PASS 'onbeforepaste' in document is true
PASS 'onpaste' in document is true
PASS 'onreset' in document is true
PASS 'onsearch' in document is true
PASS 'onselectstart' in document is true

Test Window
PASS 'onabort' in window is true
PASS 'onblur' in window is true
PASS 'onchange' in window is true
PASS 'onclick' in window is true
PASS 'oncontextmenu' in window is true
PASS 'ondblclick' in window is true
PASS 'ondrag' in window is true
PASS 'ondragend' in window is true
PASS 'ondragenter' in window is true
PASS 'ondragleave' in window is true
PASS 'ondragover' in window is true
PASS 'ondragstart' in window is true
PASS 'ondrop' in window is true
PASS 'onerror' in window is true
PASS 'onfocus' in window is true
PASS 'oninput' in window is true
PASS 'onkeydown' in window is true
PASS 'onkeypress' in window is true
PASS 'onkeyup' in window is true
PASS 'onload' in window is true
PASS 'onmousedown' in window is true
PASS 'onmousemove' in window is true
PASS 'onmouseout' in window is true
PASS 'onmouseover' in window is true
PASS 'onmouseup' in window is true
PASS 'onmousewheel' in window is true
PASS 'onscroll' in window is true
PASS 'onselect' in window is true
PASS 'onsubmit' in window is true
FAIL 'onbeforecut' in window should be true. Was false.
FAIL 'oncut' in window should be true. Was false.
FAIL 'onbeforecopy' in window should be true. Was false.
FAIL 'oncopy' in window should be true. Was false.
FAIL 'onbeforepaste' in window should be true. Was false.
FAIL 'onpaste' in window should be true. Was false.
PASS 'onreset' in window is true
PASS 'onsearch' in window is true
PASS 'onselectstart' in window is true

Test Text Node
PASS 'onabort' in textNode is false
PASS 'onblur' in textNode is false
PASS 'onchange' in textNode is false
PASS 'onclick' in textNode is false
PASS 'oncontextmenu' in textNode is false
PASS 'ondblclick' in textNode is false
PASS 'ondrag' in textNode is false
PASS 'ondragend' in textNode is false
PASS 'ondragenter' in textNode is false
PASS 'ondragleave' in textNode is false
PASS 'ondragover' in textNode is false
PASS 'ondragstart' in textNode is false
PASS 'ondrop' in textNode is false
PASS 'onerror' in textNode is false
PASS 'onfocus' in textNode is false
PASS 'oninput' in textNode is false
PASS 'onkeydown' in textNode is false
PASS 'onkeypress' in textNode is false
PASS 'onkeyup' in textNode is false
PASS 'onload' in textNode is false
PASS 'onmousedown' in textNode is false
PASS 'onmousemove' in textNode is false
PASS 'onmouseout' in textNode is false
PASS 'onmouseover' in textNode is false
PASS 'onmouseup' in textNode is false
PASS 'onmousewheel' in textNode is false
PASS 'onscroll' in textNode is false
PASS 'onselect' in textNode is false
PASS 'onsubmit' in textNode is false
PASS 'onbeforecut' in textNode is false
PASS 'oncut' in textNode is false
PASS 'onbeforecopy' in textNode is false
PASS 'oncopy' in textNode is false
PASS 'onbeforepaste' in textNode is false
PASS 'onpaste' in textNode is false
PASS 'onreset' in textNode is false
PASS 'onsearch' in textNode is false
PASS 'onselectstart' in textNode is false

Test Attribute
PASS 'onabort' in attribute is false
PASS 'onblur' in attribute is false
PASS 'onchange' in attribute is false
PASS 'onclick' in attribute is false
PASS 'oncontextmenu' in attribute is false
PASS 'ondblclick' in attribute is false
PASS 'ondrag' in attribute is false
PASS 'ondragend' in attribute is false
PASS 'ondragenter' in attribute is false
PASS 'ondragleave' in attribute is false
PASS 'ondragover' in attribute is false
PASS 'ondragstart' in attribute is false
PASS 'ondrop' in attribute is false
PASS 'onerror' in attribute is false
PASS 'onfocus' in attribute is false
PASS 'oninput' in attribute is false
PASS 'onkeydown' in attribute is false
PASS 'onkeypress' in attribute is false
PASS 'onkeyup' in attribute is false
PASS 'onload' in attribute is false
PASS 'onmousedown' in attribute is false
PASS 'onmousemove' in attribute is false
PASS 'onmouseout' in attribute is false
PASS 'onmouseover' in attribute is false
PASS 'onmouseup' in attribute is false
PASS 'onmousewheel' in attribute is false
PASS 'onscroll' in attribute is false
PASS 'onselect' in attribute is false
PASS 'onsubmit' in attribute is false
PASS 'onbeforecut' in attribute is false
PASS 'oncut' in attribute is false
PASS 'onbeforecopy' in attribute is false
PASS 'oncopy' in attribute is false
PASS 'onbeforepaste' in attribute is false
PASS 'onpaste' in attribute is false
PASS 'onreset' in attribute is false
PASS 'onsearch' in attribute is false
PASS 'onselectstart' in attribute is false

Test HTMLBodyElement
PASS 'onabort' in body is true
PASS 'onblur' in body is true
PASS 'onchange' in body is true
PASS 'onclick' in body is true
PASS 'oncontextmenu' in body is true
PASS 'ondblclick' in body is true
PASS 'ondrag' in body is true
PASS 'ondragend' in body is true
PASS 'ondragenter' in body is true
PASS 'ondragleave' in body is true
PASS 'ondragover' in body is true
PASS 'ondragstart' in body is true
PASS 'ondrop' in body is true
PASS 'onerror' in body is true
PASS 'onfocus' in body is true
PASS 'oninput' in body is true
PASS 'onkeydown' in body is true
PASS 'onkeypress' in body is true
PASS 'onkeyup' in body is true
PASS 'onload' in body is true
PASS 'onmousedown' in body is true
PASS 'onmousemove' in body is true
PASS 'onmouseout' in body is true
PASS 'onmouseover' in body is true
PASS 'onmouseup' in body is true
PASS 'onmousewheel' in body is true
PASS 'onscroll' in body is true
PASS 'onselect' in body is true
PASS 'onsubmit' in body is true
PASS 'onbeforecut' in body is true
PASS 'oncut' in body is true
PASS 'onbeforecopy' in body is true
PASS 'oncopy' in body is true
PASS 'onbeforepaste' in body is true
PASS 'onpaste' in body is true
PASS 'onreset' in body is true
PASS 'onsearch' in body is true
PASS 'onselectstart' in body is true
PASS 'onbeforeunload' in body is true
PASS 'onmessage' in body is true
PASS 'onoffline' in body is true
PASS 'ononline' in body is true
PASS 'onresize' in body is true
PASS 'onstorage' in body is true
PASS 'onunload' in body is true
PASS 'onblur' in body is true
PASS 'onerror' in body is true
PASS 'onfocus' in body is true
PASS 'onload' in body is true

Test HTMLFrameSetElement
PASS 'onabort' in frameSet is true
PASS 'onblur' in frameSet is true
PASS 'onchange' in frameSet is true
PASS 'onclick' in frameSet is true
PASS 'oncontextmenu' in frameSet is true
PASS 'ondblclick' in frameSet is true
PASS 'ondrag' in frameSet is true
PASS 'ondragend' in frameSet is true
PASS 'ondragenter' in frameSet is true
PASS 'ondragleave' in frameSet is true
PASS 'ondragover' in frameSet is true
PASS 'ondragstart' in frameSet is true
PASS 'ondrop' in frameSet is true
PASS 'onerror' in frameSet is true
PASS 'onfocus' in frameSet is true
PASS 'oninput' in frameSet is true
PASS 'onkeydown' in frameSet is true
PASS 'onkeypress' in frameSet is true
PASS 'onkeyup' in frameSet is true
PASS 'onload' in frameSet is true
PASS 'onmousedown' in frameSet is true
PASS 'onmousemove' in frameSet is true
PASS 'onmouseout' in frameSet is true
PASS 'onmouseover' in frameSet is true
PASS 'onmouseup' in frameSet is true
PASS 'onmousewheel' in frameSet is true
PASS 'onscroll' in frameSet is true
PASS 'onselect' in frameSet is true
PASS 'onsubmit' in frameSet is true
PASS 'onbeforecut' in frameSet is true
PASS 'oncut' in frameSet is true
PASS 'onbeforecopy' in frameSet is true
PASS 'oncopy' in frameSet is true
PASS 'onbeforepaste' in frameSet is true
PASS 'onpaste' in frameSet is true
PASS 'onreset' in frameSet is true
PASS 'onsearch' in frameSet is true
PASS 'onselectstart' in frameSet is true
PASS 'onbeforeunload' in frameSet is true
PASS 'onmessage' in frameSet is true
PASS 'onoffline' in frameSet is true
PASS 'ononline' in frameSet is true
PASS 'onresize' in frameSet is true
PASS 'onstorage' in frameSet is true
PASS 'onunload' in frameSet is true
PASS 'onblur' in frameSet is true
PASS 'onerror' in frameSet is true
PASS 'onfocus' in frameSet is true
PASS 'onload' in frameSet is true

Test SVGPathElement
PASS 'onabort' in pathelement is true
PASS 'onblur' in pathelement is true
PASS 'onchange' in pathelement is true
PASS 'onclick' in pathelement is true
PASS 'oncontextmenu' in pathelement is true
PASS 'ondblclick' in pathelement is true
PASS 'ondrag' in pathelement is true
PASS 'ondragend' in pathelement is true
PASS 'ondragenter' in pathelement is true
PASS 'ondragleave' in pathelement is true
PASS 'ondragover' in pathelement is true
PASS 'ondragstart' in pathelement is true
PASS 'ondrop' in pathelement is true
PASS 'onerror' in pathelement is true
PASS 'onfocus' in pathelement is true
PASS 'oninput' in pathelement is true
PASS 'onkeydown' in pathelement is true
PASS 'onkeypress' in pathelement is true
PASS 'onkeyup' in pathelement is true
PASS 'onload' in pathelement is true
PASS 'onmousedown' in pathelement is true
PASS 'onmousemove' in pathelement is true
PASS 'onmouseout' in pathelement is true
PASS 'onmouseover' in pathelement is true
PASS 'onmouseup' in pathelement is true
PASS 'onmousewheel' in pathelement is true
PASS 'onscroll' in pathelement is true
PASS 'onselect' in pathelement is true
PASS 'onsubmit' in pathelement is true
PASS 'onbeforecut' in pathelement is true
PASS 'oncut' in pathelement is true
PASS 'onbeforecopy' in pathelement is true
PASS 'oncopy' in pathelement is true
PASS 'onbeforepaste' in pathelement is true
PASS 'onpaste' in pathelement is true
PASS 'onreset' in pathelement is true
PASS 'onsearch' in pathelement is true
PASS 'onselectstart' in pathelement is true
PASS 'onbegin' in pathelement is false
PASS 'onend' in pathelement is false
PASS 'onrepeat' in pathelement is false

Test SVGSVGElement
PASS 'onabort' in svgelement is true
PASS 'onblur' in svgelement is true
PASS 'onchange' in svgelement is true
PASS 'onclick' in svgelement is true
PASS 'oncontextmenu' in svgelement is true
PASS 'ondblclick' in svgelement is true
PASS 'ondrag' in svgelement is true
PASS 'ondragend' in svgelement is true
PASS 'ondragenter' in svgelement is true
PASS 'ondragleave' in svgelement is true
PASS 'ondragover' in svgelement is true
PASS 'ondragstart' in svgelement is true
PASS 'ondrop' in svgelement is true
PASS 'onerror' in svgelement is true
PASS 'onfocus' in svgelement is true
PASS 'oninput' in svgelement is true
PASS 'onkeydown' in svgelement is true
PASS 'onkeypress' in svgelement is true
PASS 'onkeyup' in svgelement is true
PASS 'onload' in svgelement is true
PASS 'onmousedown' in svgelement is true
PASS 'onmousemove' in svgelement is true
PASS 'onmouseout' in svgelement is true
PASS 'onmouseover' in svgelement is true
PASS 'onmouseup' in svgelement is true
PASS 'onmousewheel' in svgelement is true
PASS 'onscroll' in svgelement is true
PASS 'onselect' in svgelement is true
PASS 'onsubmit' in svgelement is true
PASS 'onbeforecut' in svgelement is true
PASS 'oncut' in svgelement is true
PASS 'onbeforecopy' in svgelement is true
PASS 'oncopy' in svgelement is true
PASS 'onbeforepaste' in svgelement is true
PASS 'onpaste' in svgelement is true
PASS 'onreset' in svgelement is true
PASS 'onsearch' in svgelement is true
PASS 'onselectstart' in svgelement is true
PASS 'onbegin' in svgelement is false
PASS 'onend' in svgelement is false
PASS 'onrepeat' in svgelement is false

Test SVGAnimationElement
PASS 'onabort' in animateelement is true
PASS 'onblur' in animateelement is true
PASS 'onchange' in animateelement is true
PASS 'onclick' in animateelement is true
PASS 'oncontextmenu' in animateelement is true
PASS 'ondblclick' in animateelement is true
PASS 'ondrag' in animateelement is true
PASS 'ondragend' in animateelement is true
PASS 'ondragenter' in animateelement is true
PASS 'ondragleave' in animateelement is true
PASS 'ondragover' in animateelement is true
PASS 'ondragstart' in animateelement is true
PASS 'ondrop' in animateelement is true
PASS 'onerror' in animateelement is true
PASS 'onfocus' in animateelement is true
PASS 'oninput' in animateelement is true
PASS 'onkeydown' in animateelement is true
PASS 'onkeypress' in animateelement is true
PASS 'onkeyup' in animateelement is true
PASS 'onload' in animateelement is true
PASS 'onmousedown' in animateelement is true
PASS 'onmousemove' in animateelement is true
PASS 'onmouseout' in animateelement is true
PASS 'onmouseover' in animateelement is true
PASS 'onmouseup' in animateelement is true
PASS 'onmousewheel' in animateelement is true
PASS 'onscroll' in animateelement is true
PASS 'onselect' in animateelement is true
PASS 'onsubmit' in animateelement is true
PASS 'onbeforecut' in animateelement is true
PASS 'oncut' in animateelement is true
PASS 'onbeforecopy' in animateelement is true
PASS 'oncopy' in animateelement is true
PASS 'onbeforepaste' in animateelement is true
PASS 'onpaste' in animateelement is true
PASS 'onreset' in animateelement is true
PASS 'onsearch' in animateelement is true
PASS 'onselectstart' in animateelement is true
PASS 'onbegin' in animateelement is true
PASS 'onend' in animateelement is true
PASS 'onrepeat' in animateelement is true
PASS successfullyParsed is true

TEST COMPLETE
